Rough implementation of fhe.encrypt
This currentley allows lists or lists of lists to be used and automates
encryption somewhat, such that if the required parameters are
unspecified it will still attempt to encrypt. This follows the order of,
arguments > class state > generation , such that provided args take
precedence over using internal class self/state, then and only then if
none of the others result in anything usefull it falls back to
generating the respective requirements as needed. E.G if the public key
does not exist it will go ahead and generate the keys.