Started implementing vector class
Considering creating a 3rd or changing the name of Fhe as it handles too
much currentley. But struggling to make sense of splitting up Fhe as it
does not match conceptially. All Fhe tools should belong in Fhe and it
should encapsulate all of seal on its own. BUUUT passing itself into
itself is a bit disgusting if we were say to define the encryption
scheme.