Framed basic evector function definition
Now need to implement them, they should be a small number and very
specific, which it looks like they are. The only iffy this is whethere
to start (__init__) encrypted or to use an explicit decryptor call. For
now I am leaning on the side of explicit so the routine is pretty much
outlined, and things can be re-encrypted in place.