Added multiple key properties and workarounds
Added workaround for context requirement in set and getstate for
seal.load for certain keys. Added multiple properties to handle said
keys, including key generation in such a way as to make it as easy as
possible to functionally use this abstraction, like automatic generation
of necessary components. I.E if you want a private_key generated but
have only provided encryption parameters, it will generate the context
first, then will generate all the keys availiable from a given key
generator. I have not yet handled BFV but the general gist is there to
be expande dupon, and I have risen errors where appropriate to inform
the user that these keys are unaviliable or x generator is not there.