Added base structure for mult and add
r + 2 now works correctly adding 2 to the ciphertext returning a
ciphertext. This should begin to make it easier to operate on encrypted
vectors, but still some work to do on multiplication and handling
ciphertext + ciphertext and ciphertext * ciphertext. Will also need to
make it a tad cleaner to return ciphertexts to reseal objects (r)
instead of r.ciphertext = r + 2, to something like r = r + 2 and r += 2.