Added duplicate function
This is to optimise/ prevent the need to deep copy, by copying all by
reference except for ciphertext and cache since they are the only things
that should be different for the same key/ parameters, so lets say you
desire to reseal2 = reseal + 2, both reseal and reseal2 should keep the
same parameters and keys, since they are both born of the same, and if
they had different keys the computation would be undecryptable/ useless.
However we want reseal to continue to exist lets say for a subsequent
computation later on that is then added to the intermediary reseal2.