Added __array__ and tests to check pickle
Can now convert back to numpy array for testing, along with implementing
these checks now that we have a legible array. Added test for pickling
to ensure that it continues to work for serialisation but there is some
serious optimisation that can happen here when each encrypted vector
shares much of the same data like keys which are usually quite sizable,
in theorey all they actually need to store independantly is the
cyphertext itself assuming the shared data is stored somewhere
individually.