Added disgusting cloning method
Cloning params by passing in to constructor. Which is fine, but its kind
of messy atm so needs straightening out so it is clearer what is going
on. As atm they see clone and have absoluteley no clue what that does.
It may be neccessary to .duplicate from origin than instantiate a new
class.