Inherited Layer to CC class for brevity
Layer has now been inherited by Cross_Correlation to make uniform the
interface; being more consistent. This way the caching of intermediaries
is achieved for free. The only change that needs to happen is that bias
now needs to be moved into CNN specifically as it makes little sense to
bundle that into CC as a seperate layer now.