Crystalizing before changing activation output
activations should output shapes of (M, B) where M is the total number
of inputs that have been appended to the self.x of activations. B is the
number of batches. So if sigmoid is given 25 inputs to activate, then
M=25, and if each input has a batch size of 64 then B=64. This commit is
not the one that includes this change, but instead signposts it.