Completed sigmoid backprop in theorey
Sigmoid backprop when given the gradient ahead and the cached x will now
produce the gradient with respect to the output again. This time I have
explicitly made it so that one single instance is passed to sigmoid
only, as it is always a map function, wheras the layers themselves are
not map functions but will instead vary depending on the layer used, E.G
a CNN will reduce going backwards since we commute the sum until later.