Check deterministic loss decrease
If forward + backward + weight update are all working correctly, and 1
is the target value which we always use then the network will gradually
learn to just always predict 1. so even if it starts at 5 it will
gradually decrease down the gradients untill it hits 1. What we have now
done is added a check to find if this is the case, if the loss has
increased then we know we have done something wrong in this very
deterministic scenario since the input data is just repeatedley
forwarded every time.