Using sigmoid approximation derivative now
Instead of using standard sigmoid derivative we now use the
approximations derivative. They are both almost equal with the exception
of the asymptote outside the range -4<x>4 so clearly some activations
were still larger than these values causing the gradient to effectiveley
be flat, meaning there was no good way to go, wheras now we have a much
more stable gradient that shows us which way we need to go. However it
looks a bit strange as it is never positive outside of an incredibly
small range.