sigmoid truth inaccurate + shortened ann test
Sigmoid truth using standard numpy functions is actually inacurate on
large array since it is only built for singular values, since
broadcasting puts the values out of whack. ANN test to 50 was a bit
long, so shortened it to make it take less time. I dont think going all
the way to 50 is that valuable anyway, as long as we can see the
gradients descending determinictically thats all we care for