Removed nn branches, fixed weight init, fnd GD bug
Removed the neural network branches as in actuallity what we are
summing is the product of all the weights, so instead we init our
weights between 0-1 then divide by the numbr of weights, thus keeping
the sum total of all possible products to cap out at 1. This massiveley
solved the golden zone issue atleast initially as the weights wont
spiral to invinity instantly uppon weight init.

However solving this above issue has revealed another issue in either
gradient descent or weight update. GD seems to not differentiate between
negative and positive numbers properly, I.E is not going backwards when
it overshoots the minima. I will need to check all values are being
passed properly, and potentially check gradients manually to find out if
its a backprop or a weight-update problem.