Large strides big filters, minimise cyphertexts
In the jupyter notebook example for fashion mnist I have adapted the
strides and kernels to be quite wide and quite far apart to minimise the
amount of intermediate cyphertexts that need be stored. That way the RAM
usage falls by orders of magnitude to become reasonable again. I will
need to improve on space efficiency when possible.