Moved pages to staging for container avalability
So I realised that it was not viable to have pages on the dev branch
since the dev branch does not produce docker images which the pages are
built from. However I would still like to have as up-to date autodocs as
possible so I have opted instead to use the staging branch, and this
allows me to fiddle around with it somewhat without concern for the main
branch being effected.