Added default optimiser
Added default optimiser structure so that this could be change don the
fly. This currentley links to SGD but in future this will default to
adam. This currentley lets the optimiser self initialise its own
defaults. This is part of nodes since it is so deeply connected to the
operations of any given node that needs to update its weights/ other
parameters.