Added weight and bias to linear
Linear activation was by default identity, but now I have allowed it to
be trained with weights and biases. It still defaults to identity with
weight of 1 and bias of 0. So its now both an activation and a layer I
guess since its more than a standard linear activation now.