Added lossess as their own type of class
Losses are now class types with unique args and behaviours since they
are slightly different to the norm requiring multiple inputs. I think in
future I will be able to handle this with dictionaries so that an
abstraction can sit in front of all nodes that can handle multiple types
of inputs etc.