Moved updater to nodes abstraction
Updater is now a utility function to make it easier to update parameters
with little reused code. The only way to make this even slimmer is to
actually mark/ list the parameters using some function decorators, then
you would not even need the update/ updates functions at all and that
could be completeley handled by the abstraction.