Added more guards for stride needing to be ints
It doesnt matter so much what the meta container is around the stride,
however it must be integers! It doesnt make sense to have partial
strides, so I have added more checks to ensure it is integers, and to
automatically default to 1,1,1,1,1,1... depending on the number of
dimensions of the weights, since our stride handler can stride
selectiveley across multidimensional space.