Outlined forward pass
I am considering scrapping a "CNN" abstraction and instead using the
individual constituants of a CNN almost seperateley. This will in
particular allow us to handle the incredibly awkward sum, which can sort
of be commuted but also cant! This way I can treat the sum as a seperate
step! Could also be usefull in other contexts too.