Renamed and moved CNN to operations.CC
Since the CNN is not really a CNN since it only calculates a biased
cross correlation. So I have changed its name to CC and instead I will
add new operations that do: decryption, summation  and encryption. This
will allow me to create a CNN as nodes, while being able to interject to
handle some of the complexity involved.