Fixed firing and probe_shape being older version
Accidentally used the older version of probe shape which was missing
handling of ints/ floats. It is now using the newest version from the
CNN before it was deleted. Fixed firing and prefab although the last and
final case it will need to handle is the loss function with TWO args
instead of the usual ONE! I may just forgo the second arg and just rely
on the automatic enqueue of the firing func.