Completed firing tests
OFC we should be careful of passing by reference which is done almost
universally throught the computational graph. This helps us keep the
memory usage down but also could cause backwards-mutation if we arent
especially careful. However due to the nature of numpy maths as long as
we arent assigning back to the original variable this should not cause
us any harm.