Added fixed numpy fields to CC serialisation
Cross correlation was not serialisable with marshmallow by default so we
implemented a NumpyField custom marshmallow field. We completed the
field in the last commit and this one roles in its functionality to CC
and it properly tests the dictionaries match from the copied from, and
the copied to object after serialisation. This also overrides the
behavious of setstate and getstate to use this serialisation. This also
massiveley improves the security of the objects so when recieving them
from clients it is possible to validate them, AND ensure we are
disregarding anything potentially malitious into the void of "fob off".