Restructured examples so they each have a page
Each example is now on a seperate page for two reasons, firstly so that
the embeded HTML does not interfear with each other as the graph
visualisation can be quite tempermental. Second its a lot cleaner to
have the examples laid out like the rest of the documentation. I have
also added a preview of the constelation network, but this is still
heavily in progress.