Added selector for selecting betwene train-infer
It is necessary to add some sort of bridge between training and
inference to allow for both to co-exist in the same graph without having
to modify the graph. Thus I created the selector which can select what
goes forward and what goes backward. A nice trick is you can make
everything go forward (thus getting both inference and training) and you
can set backward to only take input from the training so then you dont
need to worry about passing back a true inference value to allow the
graph traversal algorithm to work, you can simply add a 0 or whatever,
and it will just be ignored. It just cant be None or else the graph will
think this node is not ready for computation (See neuronal firing)