Added debug in firing + improved orbweaver
Orbweaver needed some new nodes to work properly with backprop as some
of the nodes werent compatible together and needed an intermediary layer
like the selector nodes. I have also added some basic debug
functionality in firing so that we can see where the graph traversal
goes and what it sees which makes debuging 100000% times easier when you
need it. In future I also hope to be able to plot the graph, inputs
outputs everything up untill the point of failure so that it could be
closer inspected AND to even tell you which node and display the error
message because that was the entire point of graphs was to make this
kind of thing a lot easier to see, debug, test, and inevitable SOLVE