Fixed dequeue now always assuming multi forwards
backwards assume it will always get multiple sources so then summing
over axis 0 will always sum between those inputs. The only time
backwards would fail is if it only recieved a singular input but if that
is the case then there would be no point in a forward dequeue so it
should be removed. (the only special case is of course automation/
scaling which may very well lead to a singular output but still a bit of
an edge case)