Added saved data from wip implementation
Fashion-MNIST model outcomes are now recorded, we just need to
potentially add some more features, or save the models themselves, then
plot these. I have added this small data to the git repo as it is also a
good comparison for unit tests in future, to ensure results are within
the expected bounds.