Added rotation node
This node is quite special, it basically does encryption, decryption,
rotation, and plaintext passthrough all in one. It requires that the
encryptor/ provider it is given is callable, and in that case it can
copy the cyphertext objects without the data and do all sorts of neat
things with it. I will desperateley need to work on the erray
abstraction, as it will solidify the API and make it just that much
easier to use and portable without having to hard code ReArray or some
specific implementation.