Added autoFHE skip to blank nodes and cyph MNIST
Cyphertext example for fashion-MNIST is there, but I will need to hop on
over to a different machine for this, erm, VERY intensive processing!
AutoFHE now skips nodes that dont have their sources labeled as it is
assumed these are underisables. Now to check how much each node actually
costs!