Added rotate in CNN to minimise product time
Calculating the CC over a cyphertext can be quite lengthy and VERY
sizeable so we want to keep this cyphertext as small as possible. The
other cyphertexts can afford to be larger since they are A: connected
via less edges, and B: arent quite so expensive operations.

This new rotate will need some adjustments to be able to fit in here
however, as the cyphertext will need to be summed on the last axis, and
flattened ready for encryption in a flat, addable manner.